Remove cmp.ts
