+++ paf-mvp-demo-express/.gitignore
@@ -3,9 +3,6 @@
 /public/assets/app.bundle.js
 /public/assets/app.bundle.css
 /public/assets/images/
-/public/assets/cmp/cmp.js
-/public/assets/cmp/**/*.d.ts
-/public/assets/cmp/**/*.js.map
 
 # Ignore generated SSL certificates
 /paf.crt
