+++ paf-mvp-demo-express/package.json
@@ -4,8 +4,6 @@
   "main": "src/index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build-cmp": "rollup -c",
-    "build": "npm run build-cmp",
     "start": "ts-node -r tsconfig-paths/register ./src/index.ts",
     "generate-examples": "ts-node -r tsconfig-paths/register scripts/generate-examples.ts"
   },
