--- paf-mvp-demo-express/src/cmp/frontend/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "declarationDir": ""
-  }
-}
