+++ paf-mvp-frontend/src/containers/notification/Notification.tsx
@@ -23,7 +23,6 @@ export const Notification = ({ type, destroy }: INotificationProps) => {
 
   const launchPrompt = (event: MouseEvent) => {
     event.preventDefault();
-    // TODO: launch paf-lib function with callback
     window.PAFUI.promptConsent();
     window.clearTimeout(timerRef.current);
     destroy();
