+++ paf-mvp-frontend/src/lib/paf-lib.ts
@@ -488,6 +488,68 @@ export const createSeed = async (
   return await response.json();
 };
 
+/**
+ * Check existing PAF cookies and show the prompt if needed
+ * @param proxyHostName
+ * @param forcePrompt
+ */
+export const cmpCheck = async ({ proxyHostName, forcePrompt }: { proxyHostName: string; forcePrompt?: boolean }) => {
+  if (forcePrompt) {
+    removeCookie(Cookies.identifiers);
+    removeCookie(Cookies.preferences);
+  }
+
+  const { status, data } = await refreshIdsAndPreferences({ proxyHostName, triggerRedirectIfNeeded: true });
+
+  if (status === PafStatus.REDIRECT_NEEDED || status === PafStatus.NOT_PARTICIPATING) {
+    // Will trigger a redirect, nothing more to do
+    return;
+  }
+
+  const returnedId = data.identifiers?.[0];
+  const hasPersistedId = returnedId && returnedId.persisted !== false;
+
+  if (!hasPersistedId || data.preferences === undefined) {
+    // Reset cookies before to show the prompt to make sure we set the appropriate value only after user action
+    removeCookie(Cookies.identifiers);
+    removeCookie(Cookies.preferences);
+    removeCookie(Cookies.lastRefresh);
+
+    const optIn = await PAFUI.promptConsent();
+
+    if (optIn === undefined) {
+      // User closed the prompt consent without defining their preferences
+      // => means they are not participating
+      saveCookieValue(Cookies.identifiers, undefined);
+      saveCookieValue(Cookies.preferences, undefined);
+    } else {
+      // 1. sign preferences
+      const unsignedPreferences = {
+        version: '0.1',
+        data: {
+          use_browsing_for_personalization: optIn,
+        },
+      };
+      const signedPreferences = await signPreferences(
+        { proxyHostName },
+        {
+          identifiers: data.identifiers,
+          unsignedPreferences,
+        }
+      );
+
+      // 2. write
+      await writeIdsAndPref(
+        { proxyHostName },
+        {
+          identifiers: data.identifiers,
+          preferences: signedPreferences,
+        }
+      );
+    }
+  }
+};
+
 // **************************************************************************************************************** LOGS
 const label = (color: string) =>
   `display: inline-block; color: #fff; background: ${color}; padding: 1px 4px; border-radius: 3px;`;
